@@ -12,49 +12,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.4",
-    "@angular/cdk": "^7.1.0",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/material": "^7.1.0",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
-    "core-js": "^2.5.4",
-    "cors": "^2.8.5",
-    "d3": "^5.7.0",
-    "node-sass": "^4.13.1",
-    "rxjs": "~6.3.3",
-    "rxjs-compat": "^6.3.3",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "7.1.0",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/flex-layout": "7.0.0-beta.19",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "7.0.0",
+    "@angular/material": "7.1.0",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "core-js": "3.6.4",
+    "cors": "2.8.5",
+    "d3": "5.7.0",
+    "node-sass": "4.13.1",
+    "rxjs": "6.5.4",
+    "rxjs-compat": "6.3.3",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.803.23",
-    "@angular-devkit/build-ng-packagr": "~0.10.0",
-    "@angular/cli": "~7.0.4",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "~7.0.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.2.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tsickle": ">=0.29.0",
-    "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "@angular-devkit/build-ng-packagr": "0.13.5",
+    "@angular/cli": "8.3.23",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/jasmine": "2.8.8",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "8.9.4",
+    "codelyzer": "4.5.0",
+    "jasmine-core": "2.99.1",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "4.4.1",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "ng-packagr": "4.2.0",
+    "protractor": "5.4.0",
+    "ts-node": "7.0.0",
+    "tsickle": "0.29.0",
+    "tslib": "1.9.0",
+    "tslint": "5.11.0",
+    "typescript": "3.5.3"
   }
 }
